DT-337 Turn the old API test into an integration test using real data
